feat: handle downgrade install
