+++ web/app/components/plugins/plugin-detail-panel/detail-header.tsx
@@ -42,10 +42,8 @@ import { useAllToolProviders } from '@/service/use-tools'
 import DeprecationNotice from '../base/deprecation-notice'
 import { AutoUpdateLine } from '../../base/icons/src/vender/system'
 import { timeOfDayToDayjs } from '../reference-setting-modal/auto-update-setting/utils'
-import DowngradeWarningModal from '../update-plugin/downgrade-warning-modal'
 import useReferenceSetting from '../plugin-page/use-reference-setting'
 import { AUTO_UPDATE_MODE } from '../reference-setting-modal/auto-update-setting/types'
-import { useInvalidateReferenceSettings, useRemoveAutoUpgrade } from '@/service/use-plugins'
 
 const i18nPrefix = 'plugin.action'
 
@@ -132,17 +130,10 @@ const DetailHeader = ({
     return false
   }, [autoUpgradeInfo, plugin_id])
 
-  const [isShowDowngradeWarningModal, {
-    setTrue: showDowngradeWarningModal,
-    setFalse: hideDowngradeWarningModal,
-  }] = useBoolean(false)
-
+  const [isDowngrade, setIsDowngrade] = useState(false)
   const handleUpdate = async (isDowngrade?: boolean) => {
     if (isFromMarketplace) {
-      if(isAutoUpgradeEnabled && isDowngrade) {
-        showDowngradeWarningModal()
-        return
-      }
+      setIsDowngrade(!!isDowngrade)
       showUpdateModal()
       return
     }
@@ -178,20 +169,6 @@ const DetailHeader = ({
   const handleUpdatedFromMarketplace = () => {
     onUpdate()
     hideUpdateModal()
-    hideDowngradeWarningModal()
-  }
-
-  const { mutateAsync } = useRemoveAutoUpgrade()
-    const invalidateReferenceSettings = useInvalidateReferenceSettings()
-
-  const handleExcludeAndDownload = async () => {
-    await mutateAsync({
-      plugin_id,
-    })
-    invalidateReferenceSettings()
-    // TODO: missing do upgrade logic
-    handleUpdatedFromMarketplace()
-    hideDowngradeWarningModal()
   }
 
   const [isShowPluginInfo, {
@@ -376,6 +353,7 @@ const DetailHeader = ({
       {
         isShowUpdateModal && (
           <UpdateFromMarketplace
+            pluginId={plugin_id}
             payload={{
               category: detail.declaration.category,
               originalPackageInfo: {
@@ -389,16 +367,10 @@ const DetailHeader = ({
             }}
             onCancel={hideUpdateModal}
             onSave={handleUpdatedFromMarketplace}
+            isShowDowngradeWarningModal={isDowngrade && isAutoUpgradeEnabled}
           />
         )
       }
-      { isShowDowngradeWarningModal && (
-        <DowngradeWarningModal
-          onCancel={hideDowngradeWarningModal}
-          onJustDowngrade={handleUpdatedFromMarketplace}
-          onExcludeAndDowngrade={handleExcludeAndDownload}
-        />
-      )}
     </div>
   )
 }
