+++ web/app/components/plugins/update-plugin/downgrade-warning.tsx
@@ -1,5 +1,4 @@
 import { useTranslation } from 'react-i18next'
-import Modal from '@/app/components/base/modal'
 import Button from '@/app/components/base/button'
 
 const i18nPrefix = 'plugin.autoUpdate.pluginDowngradeWarning'
@@ -17,11 +16,7 @@ const DowngradeWarningModal = ({
   const { t } = useTranslation()
 
   return (
-    <Modal
-      isShow
-      onClose={() => onCancel()}
-      className='w-[640px] max-w-[640px]'
-    >
+    <>
       <div className='flex flex-col items-start gap-2 self-stretch'>
         <div className='title-2xl-semi-bold text-text-primary'>{t(`${i18nPrefix}.title`)}</div>
         <div className='system-md-regular text-text-secondary'>
@@ -33,7 +28,7 @@ const DowngradeWarningModal = ({
         <Button variant='secondary' destructive onClick={onJustDowngrade}>{t(`${i18nPrefix}.downgrade`)}</Button>
         <Button variant='primary' onClick={onExcludeAndDowngrade}>{t(`${i18nPrefix}.exclude`)}</Button>
       </div>
-    </Modal>
+    </>
   )
 }
